% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_snb}
\alias{d_snb}
\title{d_snb}
\usage{
d_snb(x, size_param = NULL, prob_param = NULL, mu_param = NULL,
  infinite = 100000L, log_v = FALSE)
}
\arguments{
\item{x}{numeric value or vector}

\item{size_param, prob_param, mu_param, size, prob, mu}{numeric value or vector,
atleast two of these parameters must be specified}

\item{infinite}{integer value specifying the maximum number
of terms to compute in the series, default: 100000}

\item{log_v, log}{whether the results should be log values,
default: false}
}
\description{
d_snb
}
